+++ raiden-dapp/src/utils/query-params.ts
@@ -1,4 +1,13 @@
 import AddressUtils from '@/utils/address-utils';
+import { BigNumber, bigNumberify } from 'ethers/utils';
+
+export function getPaymentId(queryParam: any): BigNumber | undefined {
+  try {
+    return bigNumberify(queryParam);
+  } catch (_error) {
+    return undefined;
+  }
+}
 
 export function getAmount(queryParam: any): string {
   let amount = '';